+++ man-roxygen/reference-vis-paper.R
@@ -1,3 +1,4 @@
-#' @references Gabry, J., Simpson, D., Vehtari, A., Betancourt, M., Gelman, A.
-#'   (2017). Visualization in Bayesian workflow. arXiv preprint
-#'   \href{https://arxiv.org/abs/1709.01449}{arvix:1709.01449}.
+#' @references Gabry, J., Simpson, D., Vehtari, A., Betancourt, M., and Gelman,
+#'   A. (2018). Visualization in Bayesian workflow. \emph{Journal of the Royal
+#'   Statistical Society Series A}, accepted for publication. arXiv preprint:
+#'   \url{http://arxiv.org/abs/1709.01449}.